EEIP.NET/EIPClient.cs  use ushort conversion for connectionSize
Wrapper/ImplicitConnection.cs
 set AssemblyObjectClass and ConfigurationAssemblyInstanceID

Wrapper/ImplicitParams.cs
 use ushort conversion for Length

Wrapper/Profile.cs
 set AssemblyObjectClass and ConfigurationAssemblyInstanceID

+++ Wrapper/ImplicitParams.cs
@@ -8,7 +8,7 @@ namespace Wrapper
 	public class ImplicitParams
 	{
 		public byte Instance_ID { get; set; }
-		public byte Length { get; set; }
+		public ushort Length { get; set; }
 		public ConnectionType ConnectionType { get; set; } = ConnectionType.Point_to_Point;
 		public bool OwnerRedundant { get; set; } = true;
 	}
